Revert "Milestone2 : To Do list interactive list"
