+++ modules/taskFunction.js
@@ -1,11 +1,9 @@
+import createIcon from './fontAwesome.js';
 // eslint-disable-next-line
-import { arrayOfTasks, clearItem, deleteTask } from './addItem.js';
-import checker from './checker.js';
-import createIcon from './createIcon.js';
+import {
+  taskInput, toDoList, arrayOfTasks, deleteTask,
+} from '../src/index.js';
 
-const clearAllBtn = document.querySelector('.clearList');
-
-// createElement function
 const createElement = (todo) => {
   const div = document.createElement('div');
   div.classList.add('listItem');
@@ -15,6 +13,10 @@ const createElement = (todo) => {
   //   creating div for checkbox and input
   const divElement = document.createElement('div');
   divElement.className = 'checkText';
+  // creating checkbox
+  const inputElement = document.createElement('input');
+  inputElement.type = 'checkbox';
+  inputElement.id = 'myCheckbox';
   //   creating text input
   const inputTextElement = document.createElement('input');
   inputTextElement.classList.add('toDoText');
@@ -26,28 +28,41 @@ const createElement = (todo) => {
     todo.description = newInputValue;
     localStorage.setItem('arrayOfTasks', JSON.stringify(arrayOfTasks));
   });
-  // creating checkbox
-  const inputElement = document.createElement('input');
-  inputElement.type = 'checkbox';
-  inputElement.id = 'myCheckbox';
-  //   checkbox event listener
-  inputElement.addEventListener('change', () => {
-    checker(inputTextElement, inputElement, div, todo);
-    localStorage.setItem('arrayOfTasks', JSON.stringify(arrayOfTasks));
-  });
-  //    clear all event listener
-  clearAllBtn.addEventListener('click', clearItem);
   //   appending checkbox and text input to the div
   divElement.appendChild(inputElement);
   divElement.appendChild(inputTextElement);
   // create icon function called
   const icon = createIcon('fa-trash-can');
   // add event listener delete icon
-  icon.addEventListener('click', () => deleteTask(todo.index));
+  icon.addEventListener('click', () => {
+    deleteTask(todo.index);
+  });
   //   buttonElement.appendChild(iElement);
   div.appendChild(divElement);
   div.appendChild(icon);
   return div;
 };
 
-export default createElement;
+export const displayItem = () => {
+  toDoList.innerHTML = '';
+  arrayOfTasks.forEach((todo) => {
+    const element = createElement(todo);
+    toDoList.appendChild(element);
+  });
+  taskInput.value = '';
+};
+
+export const addItem = () => {
+  if (taskInput.value !== '') {
+    const task = {
+      description: taskInput.value,
+      completed: false,
+      index: arrayOfTasks.length + 1,
+    };
+    arrayOfTasks.push(task);
+    localStorage.setItem('arrayOfTasks', JSON.stringify(arrayOfTasks));
+    displayItem();
+  }
+};
+
+export { createElement };
